added last fixes
